+++ supabase/migrations/20260911220406_allow_current_infinitepay_checkout_host.sql
@@ -1,0 +1,9 @@
+alter table public.order_payments
+  drop constraint order_payments_checkout_url_host;
+
+alter table public.order_payments
+  add constraint order_payments_checkout_url_host check (
+    checkout_url is null
+    or checkout_url ~ '^https://checkout[.]infinitepay[.]com[.]br/.+'
+    or checkout_url ~ '^https://checkout[.]infinitepay[.]io/.+'
+  );